Prevent a crash.
A battery percent report was commonly coming in immediately for some
reason, and the data structure wouldn't be ready yet, just toss away any
reports before the schema is in place.  This could potentially be made
to work but it's probably not worth it as this information will come in
right after anyway.

Also make the output a bit more useful.