Update readme to reflect project status.
I also included all the final changes, which as far as I know don't
work, or maybe appeared not to work because the device may have been
dying as I was testing them.  If someone else is interested, there's
enough information here to query the guitar and interact with its
parameters.